♻️ (#118) 탭 컨텐츠 컴포넌트로 분리
