+++ app/member/mypage/page.module.scss
@@ -36,10 +36,10 @@
 }
 
 .modalContent {
+  position: relative; // 추가
   background: white;
   padding: 20px;
   border-radius: 8px;
-  position: relative;
   width: 400px;
   max-width: 80%;
   box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.3);
